Working on alignment
This code should let the driver hit the left bumper to inch left when in front of the reef apriltags and hit the right bumper to inch right. They can also hit a to try to align with the reef using the limelight while in teleop. Ignore this class's version of scoring logic.